requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pandas==2.0.3
requests==2.31.0
streamlit==1.25.0
playwright==1.41.1
python-dotenv==1.0.0
